feat: refactor authentication flows by consolidating service functions and improving error handling
